Separated Android test code from source.
