+++ rxjava-contrib/rxjava-android/src/main/java/rx/android/schedulers/HandlerThreadScheduler.java
@@ -17,22 +17,12 @@ package rx.android.schedulers;
 
 import android.os.Handler;
 
-import org.junit.Test;
-import org.junit.runner.RunWith;
-import org.mockito.ArgumentCaptor;
-import org.robolectric.RobolectricTestRunner;
-import org.robolectric.annotation.Config;
-
 import rx.Scheduler;
 import rx.Subscription;
 import rx.operators.SafeObservableSubscription;
 import rx.util.functions.Func2;
 
 import java.util.concurrent.TimeUnit;
-
-import static org.mockito.Matchers.eq;
-import static org.mockito.Mockito.mock;
-import static org.mockito.Mockito.verify;
 
 /**
  * Schedules actions to run on an Android Handler thread.
@@ -83,49 +73,6 @@ public class HandlerThreadScheduler extends Scheduler {
             }
         }, unit.toMillis(delayTime));
         return subscription;
-    }
-
-    @RunWith(RobolectricTestRunner.class)
-    @Config(manifest=Config.NONE)
-    public static final class UnitTest {
-
-        @Test
-        public void shouldScheduleImmediateActionOnHandlerThread() {
-            final Handler handler = mock(Handler.class);
-            final Object state = new Object();
-            @SuppressWarnings("unchecked")
-            final Func2<Scheduler, Object, Subscription> action = mock(Func2.class);
-
-            Scheduler scheduler = new HandlerThreadScheduler(handler);
-            scheduler.schedule(state, action);
-
-            // verify that we post to the given Handler
-            ArgumentCaptor<Runnable> runnable = ArgumentCaptor.forClass(Runnable.class);
-            verify(handler).postDelayed(runnable.capture(), eq(0L));
-
-            // verify that the given handler delegates to our action
-            runnable.getValue().run();
-            verify(action).call(scheduler, state);
-        }
-
-        @Test
-        public void shouldScheduleDelayedActionOnHandlerThread() {
-            final Handler handler = mock(Handler.class);
-            final Object state = new Object();
-            @SuppressWarnings("unchecked")
-            final Func2<Scheduler, Object, Subscription> action = mock(Func2.class);
-
-            Scheduler scheduler = new HandlerThreadScheduler(handler);
-            scheduler.schedule(state, action, 1L, TimeUnit.SECONDS);
-
-            // verify that we post to the given Handler
-            ArgumentCaptor<Runnable> runnable = ArgumentCaptor.forClass(Runnable.class);
-            verify(handler).postDelayed(runnable.capture(), eq(1000L));
-
-            // verify that the given handler delegates to our action
-            runnable.getValue().run();
-            verify(action).call(scheduler, state);
-        }
     }
 }
